Removed SonarQube static code analysis
SonarQube community only allows you to scan the master branch for analysis. This is not feasible with our existing workflow because all code on master is supposed to be stable and production ready. For now, we're replacing SonarQube with Eslint, as it will help us accomplish the same goal of code quality.

Still need to find a security analysis tool (preferably CLI) for NodeJS applications.