Using LocalStorage provider to manage data store from single location
LocalStorage was not accessible from all files as a global object for some reason. This is the primary reason I refactored the code to use local storage from a single location and also utilized enums to manage object retrieval from the store.

Account roles have also been moved to enums